requirements.txt
flask
nltk
python-dotenv
openai
requests
pillow
google-generativeai
